Fix -Wtautological-constant-out-of-range-compare in serializer

+++ include/nlohmann/detail/output/serializer.hpp
@@ -919,7 +919,7 @@ class serializer
             }
         };
 
-        JSON_ASSERT(byte < utf8d.size());
+        JSON_ASSERT(static_cast<std::size_t>(byte) < utf8d.size());
         const std::uint8_t type = utf8d[byte];
 
         codep = (state != UTF8_ACCEPT)
